feat: ViewModel Json파일 연결중
