+++ android/app/src/main/java/com/example/sidedish/Products.kt
@@ -1,0 +1,16 @@
+package com.example.sidedish
+
+data class MenuData(
+    val count: Int,
+    val products: List<Products>
+)
+
+data class Products(
+    val productId: Int,
+    val productName: String,
+    val imageURL: String,
+    val description: String,
+    val price: Int,
+    val discountPrice: Int,
+    val discountName: String?
+)
