+++ android/app/src/main/res/layout/item_menu_data.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="menuData"
-            type="com.example.sidedish.MenuData" />
+            type="com.example.sidedish.Products" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -23,7 +23,7 @@
             android:layout_marginTop="8dp"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            bind:updateImage="@{menuData.representImages}" />
+            bind:updateImage="@{menuData.imageURL}" />
 
         <TextView
             android:id="@+id/tv_home_name"
@@ -89,14 +89,14 @@
             android:background="@drawable/bg_event_badge"
             app:layout_goneMarginLeft="20dp"
             android:gravity="center"
-            android:text="@{menuData.eventInfo}"
+            android:text="@{menuData.discountName}"
             android:paddingStart="12dp"
             android:paddingTop="4dp"
             android:paddingBottom="4dp"
             android:paddingEnd="12dp"
             android:textColor="#F2E7FE"
             android:textSize="12sp"
-            bind:updateVisible="@{menuData.eventInfo}"
+            bind:updateVisible="@{menuData.discountName}"
             app:layout_constraintLeft_toRightOf="@+id/iv_home_image"
             app:layout_constraintTop_toTopOf="parent" />
